lab3-leaderboard 2 finished hashjoin optimizer fixed and improved

+++ src/include/execution/executors/nested_index_join_executor.h
@@ -58,6 +58,6 @@ class NestIndexJoinExecutor : public AbstractExecutor {
 
   Tuple left_tuple;
   std::vector<Tuple> right_tuples;
-  unsigned long ptr;
+  size_t ptr;
 };
 }  // namespace bustub
